Show how to install spData in c2, fixing #52
Yes all deps should be installed with geocompr.

But, to be fair to @respite99, it does say to install the pkgs and in case you're new to R it's worth seeing how.

Also @Nowosad notice that this removes the duplication of information in the index because installing geocompr installs spData - so a little more concise there.

Thanks for the feedback - hope this rather long commit message clarifies things further.