Add function to save print quality maps
Plan: do not save large files (e.g. the results of this) in the repo to prevent 'repo bloat'

I suggest we just upload them as images to this thread so we can see and comment on them all: https://github.com/Robinlovelace/geocompr/issues/41

Function may need a couple of tweaks but seems to work.

Heads up @Jannes-m and @Nowosad and sorry for exessively long commit msg!


Former-commit-id: 0e09fcb4358dfcca75fbe3e0e425d4775e0a55e0