Suggested title change
@Jannes-m my thinking here is that a shorter and more general title will be easier to digest, especially for beginners.

I wanted to shorten the title further and add the word 'ecology' but wanted to check: how would you feel about calling it simply "Ecological applications" and using the beginning and end sections to explain how this example could be generalised? I'm thinking of doing the same thing in the transport chapter.


Former-commit-id: 25e8d7eb2fe92f8bcd271cd99557b7ee28f81367