Rewrite preface
Heads-up @Jannes-m and @Nowosad this makes it publicly acceptable (was previously unfinished) and hopefully better.

We should refer to other works for sure but this is part is now commented for now as it was not 'camera ready'. Great to have this 'who is this for' in there though. One suggestion: the first target audience should be people who use QGIS, ArcMap etc but want something more reproducible and powerful.