Geomarketing rev rl2 (#129)
* Remove library(ggmap) (called with ::)

* Don't load classint

* Reword parts of intro: more concise and hopefully clear

* Update typical location analysis questions

@Jannes-m we should decide whether to use 'geomarketing' or 'location analysis' and stick with it for consistence. I suggest the latter in this commit for consistency with the chapter and because the term 'location analysis' is less jargonistic and more-widely known.

* Fix grammar and short prose in 7.1

* Fix typo

* Reduce duplication in description of chapter

* Simplify dl code thanks to tinyurl

* Further shorten csv loading code

Seem OK @Jannes-m? Could change this further but think this updated version is easier to understand and type. Puts a few MB of ignored files into the geocompr folder.

* More rewording + making things more concise

* More tweaks to 7.1

* Make tidying more prominent in c7

By splitting it out into a new subsection

cc @Jannes-m this addresses a tickbox in #12

* Reword 7.2.2, slightly

* Update raster creation with new names

* Refactor raster creation code

Heads-up @Jannes-m this is shorter and hopefully easier to understand.

Please double check. Note new object names - safer to teach people to create objects with meaningful names rather than re-writing exiting objects is the thinking here.

* More refactoring: use purrr code

This works well with the list of rasters created using the lapply code in the previous code

* Refactor 7.2.2 code more

- Replace lapply with map and function(x) with ~

- Move the crs spec into stack creating room to delete a line

* Fix type: use Figure not Fig.

Cc @Jannes-m

* Reword description of raster creation code

* Fix broken label

* Refactor prose describing 'reclassification' in 7.2.2

@Jannes-m this is awesome stuff! Great read. Very educational. I think it should link-back to the section that introduced reclassify() - will add that now.

* Add cross-reference to 4.3.3

Heads-up @Jannes-m this is where reclassify() was introduced. Potentially useful to refer lost readers back to that if they are struggling here.

* Rename input -> input_ras, fix travis

* input_ras -> reclass

* Avoid using intermediary polys object

* Refactor parts of c7. Fix crs code.

* Use input_ras[[1]] which replaces input

Hope this fixes Travis!

* Don't load sp

* Close #113

More could be done here and I'd be interested to hear views from @Nowosad but this is clearly an excellent chapter and I'm happy with it. Great work Jannes!


Former-commit-id: d0f9a7215653523cc909068ac5ffe2e21ea72e92