Links to PROJ4 resources not working (404 error)
The links provided in line 952 do not currently open a document. I went to the PROJ4 site and tried to find the relevant reference page based on the original URL. These are pages, not HTML documents, so not sure the links I have included are the ones you seek. This section is so great, by the way. I've never read a better description of the datum and it's purpose.